requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
PyJWT==2.8.0

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1

# AI / ML
transformers==4.41.1
sentence-transformers==3.0.0

# LangChain RAG
langchain==0.2.1
langchain-community==0.2.1
faiss-cpu==1.8.0

# Frontend (if needed – this file is for backend, but having it here doesn't hurt)
streamlit==1.35.0
requests==2.32.3
pandas==2.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.7.2

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0